+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 #### Added
 - Security improvements in script calls
+* Add theme 'einfra'
 
 #### Changed
 * Updated disco-tpl.php to be compatible with the new style of configuration from module Perun v5.0